requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dependency-injector==4.46.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13